backend/requirements.txt
fastapi
uvicorn
pyyaml
pyjson
pydantic
